% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateExamSheets.R
\name{createAnswerKey}
\alias{createAnswerKey}
\title{Generate Exam Answer Key}
\usage{
createAnswerKey(numberOfQuestions, writeToFile = FALSE, moduleID,
  ansOptions = letters[1:5])
}
\arguments{
\item{numberOfQuestions}{a numerica value that specifies how many answers
should be generated for the key.}

\item{writeToFile}{if TRUE, a file, named based on the moduleID, is created
with the answer key inside.}

\item{moduleID}{a string that designates what the name of the module is. If
writeToFile = TRUE, then this arguement must be specified.}

\item{ansOptions}{the possible answers in the multiple choice exam key. The
default value, letters[1:5], specifies that there are 5 different options for
each question, a, b, c, d, and e.}
}
\value{
a vector that contains a randomized answer for each question number.
If writeToFile = TRUE, then this vector is written to a file.
}
\description{
Outputs a randomized answer key for a given exam based on
number of questions. If writeToFile = TRUE, then moduleID must be prodived to
create the file name.
}
\examples{
## create 100 question answer key
createAnswerKey(100)

## write answer to key yot file
createAnswerKey(100, writeToFile = TRUE, 'BS281')

}
\author{
Andrew Davis \email{adavisd@essex.ac.uk}
}
