% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateExamSheets.R
\name{generateAllStudentsAnswersForExam}
\alias{generateAllStudentsAnswersForExam}
\title{All Student Answers for a Given Exam}
\usage{
generateAllStudentsAnswersForExam(moduleID, numberOfQuestions = questions,
  allStudentIDs = students, examsPerSubject = exams, writeToFile = FALSE,
  readFromFiles = FALSE, degreeNames = NULL)
}
\arguments{
\item{numberOfQuestions}{the dataframe that contains the amount of questions
each student needs to answer for each exam. The defualt value, questions, is
a dataframe included in the package.}

\item{allStudentIDs}{the dataframe that contains the ID for each student and
what degree course they are on. The defualt value, students, is a dataframe
included in the package.}

\item{examsPerSubject}{the dataframe that contains a dataframe that lists
what modules a given degree course takes. The first column should list
modules, the second column should list the options for options for
degree 1, and the third column should list options for
degree 2. The possible options are "Yes", "No", and "Optional". If a string
is supplied that is not one of the three then it is evaluated as "Yes". if a
module is "Optional", then a random amount of students is picked to take the
exam, with a higher number of students being more likely. The defualt value,
exams, is a dataframe included in the package, with degree 1 being
"Biological Sciences" and degree 2 being "Genetics".}

\item{writeToFile}{if TRUE, a folder, named based on the moduleID is created
with a file for each student's answers. A file is also created that lists all
of the students that took the exam. The default value is set to FALSE.}

\item{readFromFiles}{if TRUE, filenames are used to read in data for the
relevant arguements instead of dataframes within R.}

\item{degreeNames}{if degree names are not "Biological Sciences" and
"Genetics" then a string should be entered with the two degree courses that
the student set belongs to.}

\item{molduleID}{a string identifying the name of the module.}
}
\value{
A list with 2 elements, a data frame of students that took the module
and a list of answers by each student. If writeToFile = TRUE, then files are
written instead.
}
\description{
Outputs random answers to a multiple choice exam for a given
module for all students. Answers can be from a to e, with NA's indicating the
question was not answered. If writeToFile = TRUE, then a folder is created
with all of the student answers. A file is also created with the list of
students that took a given exam. Files are created as .tsv's. If
readFromFiles = TRUE, then the arguements numberOfQuestions, allStudentIDs,
and examsPerSubject are read from files instead of from dataframes.
}
\examples{
## create answers for BS284 and output to console
generateAllStudentsAnswersForExam('BS284', writeToFile = FALSE)

## create files with student answer files and a list of students taking exam
generateAllStudentsAnswersForExam('BS284', writeToFile = TRUE)

}
\author{
Andrew Davis \email{adavisd@essex.ac.uk}
}
