% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateExamSheets.R
\name{generateStudentAnswersForExam}
\alias{generateStudentAnswersForExam}
\title{Student Answers for Specific Student}
\usage{
generateStudentAnswersForExam(totalNumberofQuestions, numberOfQuestionsToAnswer,
  writeToFile = FALSE, moduleID, studentID)
}
\arguments{
\item{totalNumberofQuestions}{the total number of questions in an exam}

\item{numberOfQuestionsToAnswer}{the amount of questions that the student is
asked}

\item{writeToFile}{if TRUE, a file is created with the dataframe inside. The
default value is set to FALSE.}

\item{studentID}{a string or value identifying the student that took the exam
Only needed if a file is being written.}

\item{molduleID}{a string identifying the name of the module. Only
needed if a file is being written.}
}
\value{
a dataframe with a question and answer column with the option to
write dataframe to a file
}
\description{
Outputs a students random answers to a multiple choice exam in
the form of a dataframe given the number of questions asked and total number
of questions available. Answers can be from a to e, with NA's indicating the
question was not answered. If writeToFile = TRUE, then studentID and moduleID
must be provided so that the appropriate file name can be created. Files are
created as .tsv's.
}
\examples{
## create answers for an exam with 100 questions and 30 questions asked
generateStudentAnswersForExam(100, 30)

## write the student's randomized answers to a file "BS281_answers_12.tsv"
generateStudentAnswersForExam(100, 30, writeToFile = TRUE,
     moduleID = 'BS281', studentID = '12')

}
\author{
Andrew Davis \email{adavisd@essex.ac.uk}
}
